+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.TeamsComplianceRecordingPolicy.Tests.ps1
@@ -63,7 +63,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description                                         = 'FakeStringValue'
                     Enabled                                             = $True
                     DisableComplianceRecordingAudioNotificationForCalls = $True
-                    ComplianceRecordingApplications                     = 'FakeStringValue'
+                    ComplianceRecordingApplications                     = @(@{Id="123456"})
                     Identity                                            = 'FakeStringValue'
                     Ensure                                              = 'Present'
                     Credential                                          = $Credential
@@ -95,7 +95,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description                                         = 'FakeStringValue'
                     Enabled                                             = $True
                     DisableComplianceRecordingAudioNotificationForCalls = $True
-                    ComplianceRecordingApplications                     = 'FakeStringValue'
+                    ComplianceRecordingApplications                     = @(@{Id='123456'})
                     Identity                                            = 'FakeStringValue'
                     Ensure                                              = 'Absent'
                     Credential                                          = $Credential
@@ -107,15 +107,27 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         Description                                         = 'FakeStringValue'
                         Enabled                                             = $True
                         DisableComplianceRecordingAudioNotificationForCalls = $True
-                        ComplianceRecordingApplications                     = @{Id='FakeStringValue'}
+                        ComplianceRecordingApplications                     = "Microsoft.Teams.Policy.Aministration.Cmdlets.Core.CompianceRecordingApplication"
                         Identity                                            = 'FakeStringValue'
-
                     }
                 }
+                Mock -CommandName Get-CsTeamsComplianceRecordingApplication  -MockWith {
+                    return @{
+                        Identity                                            = 'FakeStringValue/123456'
+                        Id                                                  = '123456'
+                    }
+                }
+
             }
 
             It 'Should return Values from the Get method' {
-                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+                $Result = (Get-TargetResource @testParams)
+                $Result.Ensure | Should -Be 'Present'
+                $Result.ComplianceRecordingApplications.Length | Should -Be 1
+                $Result.ComplianceRecordingApplications[0].Id | Should -Be '123456'
+                Should -Invoke -CommandName Get-CsTeamsComplianceRecordingPolicy -Exactly 1
+                Should -Invoke -CommandName Get-CsTeamsComplianceRecordingApplication -ParameterFilter {$Filter -eq 'FakeStringValue/*'} -Exactly 1
+
             }
 
             It 'Should return true from the Test method' {
@@ -135,7 +147,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description                                         = 'FakeStringValue'
                     Enabled                                             = $True
                     DisableComplianceRecordingAudioNotificationForCalls = $True
-                    ComplianceRecordingApplications                     = 'FakeStringValue'
+                    ComplianceRecordingApplications                     = @(@{Id='123456'})
                     Identity                                            = 'FakeStringValue'
                     Ensure                                              = 'Present'
                     Credential                                          = $Credential
@@ -147,9 +159,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         Description                                         = 'FakeStringValue'
                         Enabled                                             = $True
                         DisableComplianceRecordingAudioNotificationForCalls = $True
-                        ComplianceRecordingApplications                     = @{Id='FakeStringValue'}
+                        ComplianceRecordingApplications                     = "Microsoft.Teams.Policy.Aministration.Cmdlets.Core.CompianceRecordingApplication"
                         Identity                                            = 'FakeStringValue'
 
+                    }
+                }
+                Mock -CommandName Get-CsTeamsComplianceRecordingApplication  -MockWith {
+                    return @{
+                        Identity                                            = 'FakeStringValue/123456'
+                        Id                                                  = '123456'
                     }
                 }
             }
@@ -166,7 +184,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description                                         = 'FakeStringValue'
                     Enabled                                             = $True
                     DisableComplianceRecordingAudioNotificationForCalls = $True
-                    ComplianceRecordingApplications                     = 'FakeStringValue'
+                    ComplianceRecordingApplications                     = @{Id='123456'}
                     Identity                                            = 'FakeStringValue'
                     Ensure                                              = 'Present'
                     Credential                                          = $Credential
@@ -178,8 +196,15 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         Description                                         = 'FakeStringValueDrift' #Drift
                         Enabled                                             = $False
                         DisableComplianceRecordingAudioNotificationForCalls = $False
-                        ComplianceRecordingApplications                     = @{Id='FakeStringValueDrift'} #Drift
+                        ComplianceRecordingApplications                     = "Microsoft.Teams.Policy.Aministration.Cmdlets.Core.CompianceRecordingApplication"
                         Identity                                            = 'FakeStringValue'
+                    }
+                }
+
+                Mock -CommandName Get-CsTeamsComplianceRecordingApplication  -MockWith {
+                    return @{
+                        Identity                                            = 'FakeStringValue/123456Drift'
+                        Id                                                  = '123456Drift'  #Drift
                     }
                 }
             }
@@ -212,11 +237,18 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                         Description                                         = 'FakeStringValue'
                         Enabled                                             = $True
                         DisableComplianceRecordingAudioNotificationForCalls = $True
-                        ComplianceRecordingApplications                     = @{Id='FakeStringValue'}
+                        ComplianceRecordingApplications                     = "Microsoft.Teams.Policy.Aministration.Cmdlets.Core.CompianceRecordingApplication"
                         Identity                                            = 'FakeStringValue'
 
                     }
                 }
+                Mock -CommandName Get-CsTeamsComplianceRecordingApplication  -MockWith {
+                    return @{
+                        Identity                                            = 'FakeStringValue/123456'
+                        Id                                                  = '123456'
+                    }
+                }
+
             }
 
             It 'Should Reverse Engineer resource from the Export method' {